demoTrack - Backend (Springboot)/src/main/java/com/tcs/internship/demotrack/decryptor/Decrypter.java
package com.tcs.internship.demotrack.decryptor;

import java.util.Base64;

import javax.crypto.Cipher;
import javax.crypto.spec.IvParameterSpec;
import javax.crypto.spec.SecretKeySpec;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class Decrypter {

	private static String initVector;

	private static String key;

	@Value("${decrypter.initVector}")
	public void setInitVector(String iv) {
		initVector = iv;
	}

	@Value("${decrypter.key}")
	public void setKey(String k) {
		key = k;
	}

	@RequestMapping(path = "print")
	public String decrypt(String encrypted) {
		try {
			IvParameterSpec iv = new IvParameterSpec(initVector.getBytes("UTF-8"));
			SecretKeySpec skeySpec = new SecretKeySpec(key.getBytes("UTF-8"), "AES");
			Cipher cipher = Cipher.getInstance("AES/CBC/PKCS5PADDING");
			cipher.init(Cipher.DECRYPT_MODE, skeySpec, iv);
			byte[] original = cipher.doFinal(Base64.getDecoder().decode(encrypted));
			return new String(original);
		} catch (Exception ex) {
			return null;
		}

	}

}
